Initial Query Selector Fix
I think this was necessary because of recent work done to PouchDB
with regard to conforming PouchDB operators to CounchDB's operators.

Without this fix if I pass the following selector I get an error:

    { $regex: `^grade/${params.day}` }

It complains saying that the value given to $regex is the wrong type
and indicates it is an array. Obviously it is not but I traced through
the code and found the "massageSelector" method wraps it into an array:

https://github.com/pouchdb/pouchdb/blob/bbeb7e557658fb405b5cf91dec50a79faf5278a3/packages/node_modules/pouchdb-selector-core/src/utils.js#L229

This massageSelector method is an internal method to PouchDB that
normalizes the selector to make it easier to process by the rest of
PouchDB internals. But it is important to note that this massaged
selector is for easier coding and is not actually a selector that will
pass validation (i.e. is not a selector that should be passed to
pouchdb-find). This is highlighted in PouchDB here:

https://github.com/pouchdb/pouchdb/blob/bbeb7e557658fb405b5cf91dec50a79faf5278a3/packages/node_modules/pouchdb-find/src/adapters/local/find/index.js#L77-L79

live-find uses massageSelector for much the same purpose as PouchDB
itself. To be able to easily filter new/updated records to see if it
matches the selector. This is fine for that filter purpose but the
original selector prior to the massage should be passed to pouchdb-find
so it will pass validation.